JP-2019: Added Phenomena translations and added phenomena to register type

+++ comu/com.energyict.mdc.device.config.rest/src/main/java/com/energyict/mdc/device/configuration/rest/impl/LoadProfileConfigurationResource.java
@@ -289,7 +289,7 @@ public class LoadProfileConfigurationResource {
         return channelSpec;
     }
 
-    private Phenomenon findPhenomenonByIdOrThrowException(int phenomenonId) {
+    private Phenomenon findPhenomenonByIdOrThrowException(long phenomenonId) {
         Phenomenon phenomenon = masterDataService.findPhenomenon(phenomenonId).orNull();
         if (phenomenon == null){
             throw new TranslatableApplicationException(thesaurus, MessageSeeds.NO_PHENOMENON_FOUND, phenomenonId);
